[SparkSQL][SPARK-4528] add comment support for input from a file
